[dynamo][numpy] Support ndarray methods
This PR adds universal support for ndarray methods. After #100839 each `NumpyNdarrayVariable` should wrap a `torch.Tensor`. This PR adds a `numpy_method_wrapper` which converts the `torch.Tensor` to `torch_np.ndarray` and then call the numpy ndarray method. Then we also try to return a `torch.Tensor` (return as-is if the value is not ndarray-like)

ghstack-source-id: 2805b6964d20d14a1cc37fcebc1cdebe1684f43f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97537