_cycleviz.py: visualize reference cycles holding cuda memory
Reference cycles are freed by the cycle collector rather than being cleaned up
when the objects in the cycle first become unreachable. If a cycle points to a tensor,
the CUDA memory for that tensor will not be freed until garbage collection runs.
Accumulatin of CUDA allocations can lead to out of memory errors (OOMs), as well as
non-deterministic allocation behavior which is harder to debug.

This visualizer installs a garbage collection hook to look for cycles containing
CUDA tensors and saves a visualization of the garbage:

> from torch.cuda._cycleviz import warn_tensor_cycles
> warn_tensor_cycles()
> # do some work that results in a cycle getting garbage collected
> # ...
> WARNING:root:Reference cycle includes a CUDA Tensor see visualization of cycle /tmp/tmpeideu9gl.html

ghstack-source-id: 68dc74046ce66aab9f9976297777f60970075716
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102656