[quant][pt2] Add prepare QAT test for mobilenetv2
Summary: Prepare QAT for mobilenetv2 has matching numerics with
FX. There were two changes needed to achieve this, however.
First, this commit adds observer sharing for ReLU6, which is
used extensively throughout this model. Second, in the tests we
have to use the same manual seed every time we call the models
in order to get the same results between FX and PT2. This is
because there is a dropout at the end of the model.

Test Plan:
python test/test_quantization.py TestQuantizePT2EModels.test_qat_mobilenet_v2

Reviewers: jerryzh, kimishpatel

Subscribers: jerryzh, kimishpatel