[operator_compile_check] Add FakeTensor testing
This PR adds dedicated FakeTensor testing to operator_compile_check. We
reuse CrossRefFakeMode to do this and improve the error messages on it.

Note that this only really runs detailed tests for operators that do not
have data-dependent output shape. In the future we should add something
like a dynamic CrossRefFakeMode.

Test Plan:
- existing tests (these now have improved error messages).

ghstack-source-id: e16a55e57e6c696b808ce1cabf42680f532a3fbb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103595