Cast computation_node_input_size to int
This bandaid fixes yolov3 with automatic_dynamic_shapes.
A more proper fix probably is to figure out why when we
have

```
TypeError: mkldnn_reorder_conv2d_weight(): argument 'input_size' (position 6) must be tuple of ints, but found element of type SymInt at pos 1
```

where the SymInt is known to be constant, we aren't willing to
coerce it to int.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 8a40760459bd840d663af856987d6989f06db052
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103677