[inductor] Install guards on both cases of View.handle_negative_index
This branch is not an optimization, it's a correctness issue so there should be
a guards installed for both sides of the branch. Otherwise we could have an
expression like `(s0 - s1)` that is initially positive, then becomes negative
with a new set of shapes and now references an invalid index.

ghstack-source-id: 5aac93ebfc312357eadea72c07cade3cf073d66c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103780